Added error handling functionality
If the app has an error_handler function defined, then error handler kicks in, otherwise no change.

The intent is to automatically handle errors in most cases, while allowing app code to handle errors explicitly when required. After a plugin API call, if error or truth status is checked, that clears the error state. Otherwise, ret.value call or the next plugin API call will raise an error. So if the app code explicitly checks error status, then no automatic error handling is done. If not, the error is passed to the error_handler function. The error handler function can do a retarget for partial requests and return an error page for non-HTMX requests.

For dev apps, the source location is included in the error message.